docs(README): automatic update
  Update README Scala
  [This list might not be relevant]
  Add TOC for recently added geo search parameters
Add missing `aroundRadius`, `aroundPrecision` and `minimumAroundRadius` search parameters
Reformatting and rephrasing
Ensure code blocks and lists are separated by a blank line

Some Markdown processors (such as that used by Jazzy) are very picky about that. If a list or a code block is not separated from the previous paragraph by a blank line, it gets treated as a continuation of the paragraph, resulting in the loss of formatting. [ci skip]