Fixed: Introduce 'delete objectFromIndex X' to deprecate 'delete object X from index Y'
As we recently discovered, the following request was able to delete an
entire index:

```
delete from "index" `object` ""
```

To fix this issue, we deprecated this way of deleting the object with
the following approach:

```
val op: Option[SafeDeleteObjectOperation] = SafeDeleteObjectOperation("index", "objectID")
delete objectFromIndex op.get
```

which forces the user to pass non-empty index name and objectID,
otherwise the returned `Option` is `None`. This is made possible by the
use of private constructors on the `SafeDeleteObjectDefinition` and of a
companion object with a public `apply()` method that only lets the user
instantiate a `SafeDeleteObjectDefinition` with valid i.e. non-empty
parameters.