fix: use unique ID in threads generated by the waitFor's ThreadFactory
When performing a waitFor operation, the ThreadFactory used to start new
daemon threads was always using the same name. Because our tests were
often flaky, we suspected the unique name of the generated thread to be
the issue.

This commit ensures that the thread ID are unique by appending a
timestamp to the end of each thread name.

[changelog]