added: findFirstObject() method
FindFirstObject searches iteratively through the search response `Hits`
field to find the first response hit that would match against the given
`filterFunc` function.

If no object has been found within the first result set, the function
will perform a new search operation on the next page of results, if any,
until a matching object is found or the end of results, whichever
happens first.

To prevent the iteration through pages of results, `doNotPaginate`
parameter can be set to true. This will stop the function at the end of
the first page of search results even if no object does match.

[changelog]