test: fix NullPointerException introduced by recent ScalaMock changes
Since our upgrade from ScalaMock dependency from v3.4.2 to v4.4.0
(actually, v4+), we cannot directly mock a case class with default
arguments in their constructor. Indeed, those values seem not to be
instantiated, which ends up generating a NullPointerException at the
mock instantiation call site.

The workaround is to introduce an intermediate class that extends the
case class having the default arguments in its constructor and mock the
former. This way, the default arguments are correctly instantiated.