chore: add scala-collection-compat as a new dependency to support Scala 2.13
Following up on @Ant-hem’s question on Github [^1], who was asking why
`-Xfatal-warnings` was temporarily commented.

This was removed in the original PR [^2] probably because of the deprecation
warnings that started to appear when compiling with 2.13. Because this
flag makes the compilation fail if there are warnings, I assume @dragisak
commented it to let us decide whether or not we want to keep the flag or be
OK with the warnings when compiling in 2.13.

Because we want to keep the warnings and consider them as errors, I've tried to
update the 4 deprecation warnings manually to let 2.13 compile with no warning
at all. However, by doing so, I had to use the new collections introduced by
Scala 2.13 which do not seem to be backported to Scala 2.11 and Scala 2.12 yet.

I've fiddled around finally found this new dependency which is made precisely
for this: `scala-collection-compat` [^3]. I'm now able to cross-compile the
entire project (library and tests) for all Scala versions we support with 0
warning using:

```
sbt
> +clean
> +clean-files
> +compile
> +test:compile
```

[^1]: https://github.com/algolia/algoliasearch-client-scala/pull/571#discussion_r369964488
[^2]: https://github.com/algolia/algoliasearch-client-scala/pull/553
[^3]: https://github.com/scala/scala-collection-compat