feat(rule): implement multi-condition rules
Close #591

This is a breaking change since the `Rule.condition` field type has to be
changed from `Condition` to `Option[Condition]`. IMHO it's fine to break the
compatibility here since the condition type was incorrect in the first place,
and it's more important to stay align as much as we can with other API clients.